requirements.txt
requests==2.31.0
python-dotenv==1.0.0
argparse
colorama==0.4.6
pillow==10.0.0
tqdm==4.67.1